Remove a check because it is redundant
In block_from_source, it is important to call
add_known_block_to_source_and_set_best even if the block is below the
finalized block, knowing that the
"add_known_block_to_source" part will
have no effect. However if don't update the best block of the source,
this has no reason to be done and we can check if the block is
finalized before and merge it with the check number == 0.