Refactor all forks a bit (#2129)
* Copy body of header_from_source into add_block

* Adjust copy-pasted code to parameter values

* Some more adjustments

* Merge with the block below

* Remove block and run rustfmt

* Remove a check because it is redundant

In block_from_source, it is important to call
add_known_block_to_source_and_set_best even if the block is below the
finalized block, knowing that the
"add_known_block_to_source" part will
have no effect. However if don't update the best block of the source,
this has no reason to be done and we can check if the block is
finalized before and merge it with the check number == 0.

* Split end of `add_block` into separate state machines

* Move finalized check back to add_block

* Update all.rs

* Make it compile

* Add some TODOs

* Move view update to block insertion stage

* Add AlreadyInChain to the success case

* Merge block_from_source into block_announce

* Remove unused HeaderFromSourceOutcome

* Address more unused fields

* Rename the AllBlock variants

* Documentation

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>