Provide a Host in the WebSocket handshake (#2694)
Right now, on the client side, we pass `Host: ...` in the HTTP headers.

While libp2p itself doesn't care about the value in the `Host` header, many nodes that use WebSocket have an HTTP reverse proxy in front of them. This reverse proxy does care about this value, and it is thus important to pass it properly.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>