General clean-up of the single stream connections code (#2708)
This PR is a general clean-up of the single stream connection code. It fixes many small issues, and also handles the shutdown phase more properly.
We now properly handle incoming GoAway frames, and have the possibility to send some.

Can be reviewed commit by commit.
I understand that this code is generally extremely complicated, but I don't really have a good solution for this. Networking code *is* extremely complicated, both in terms of logic and implementation.
Maybe a solution is tests, such as https://github.com/paritytech/smoldot/pull/2323, but at the moment the code isn't even robust enough to have tests.