Don't ignore extra substreams opened during handshake (#2774)
Right now, if an extra substream is opened during the handshake, we
simply ignore it. The substream will plain not work.

In principle, no extra substream should be opened during the handshake,
provided we are the client. However, due to bugs and implementation
difficulties, it is actually not impossible that this happens in
reality, in which case simply ignoring the substream is the worst thing
we can do.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>