Several clean ups and refactors of peers.rs (#2795)
This PR does a major clean-up pass on `peers.rs` and cleans up the
remnants of the old times when it was asynchronous.

In details:

- Completely remove the `DesiredOutNotificationId` system where you
first allocated a "desired outbound request" and then send the handshake
with a different function.
- Add a `unfulfilled_desired_peers` field so that we simply return the
content of the field in the `unfulfilled_desired_peers()` function
instead of iterating through everything.
- ~~When a substream is marked as "not desired" and it is currently
opening, we no longer immediately cancel the opening and instead we
generate an event about the opening having failed. This makes it
possible for API users to track which substreams are currently being
open.~~ (reverted this change, so that the PR doesn't change any of the
behavior)
- A couple of minor changes: remove a redundant function, add
documentation, etc.

It is not necessarily clear from the diff, but the logic is now way more
simple of the module, which was the objective of this PR.

I have other changes in mind, which I'll do later as they also touch
files other than `peers.rs`.