Don't stall syncing if gap is too large (#2801)
cc https://github.com/paritytech/smoldot/issues/2793

After trying to analyze the problem, I've decided to leave it to
future-me.
It is very complicated to determine which entries in the list of "blocks
known by a source" should be kept and which can be thrown away. There
are several other TODOs related to that in the code.

For now, this PR fixes the sync stalling.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>