Simplify handling of incoming notifications substreams (#2827)
This PR does two things:

- Change the API of `collection.rs` with a new
`NotificationsInOpenCancel` event. The API layer right below and the API
layer right above already have this event, so adding it to
`collection.rs` provides uniformity and removes confusion.
- Simplify the API of `peers.rs` w.r.t. incoming substreams. The API is
now more or less a wrapper around the one of `collection.rs`, the only
difference being that we refuse multiple substreams from the same
peer-protocol combination (the role of `peers.rs` is partly to enforce
the uniqueness of peer-protocol combinations).

Overall this simplifies quite a lot the logic of this code by making it
similar across all API layers, thus making it easier to understand.