Fix Yamux not picking up events if no data is coming (#2839)
At the moment, if we receive a header with `fin: true` and `length: 0`,
we do a state transition to `Incoming::DataFrame`. Then, when we loop
again, we will generate a `StreamClosed` event. This is all correct.

Unfortunately, the loop condition is `while !data.is_empty()`, meaning
that if no more data comes we will never generate the `StreamClosed`
event.
I've changed it to `loop`.

Additionally, I've cleaned up the handling of `DataFrame` and `Window`
headers, so that it is a bit more elegant.

This PR also fixes an API inconsistency: right now a substream can be
considered as dead even before we have generated the `StreamClosed`
event for it. Since we now transition the substream to "dead" right
before generating the event, there's no weirdness anymore.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>