Properly process dead substreams (#2869)
At the moment, when a substream is closed gracefully, the code in
`established::single_stream` just assumes that the substream state
machine will return zero or one event then shut down.

While this is true in practice, this is a pretty big breach of
abstraction layers. The state machine could in principle decide to
return two events. Nothing in the API prevents it from doing so. And if
it does, before this PR the second event would be ignored.

This PR splits the dead substreams handling of the yamux code in two:
first we iterate over dead substreams, then only we remove them. Thanks
to this, we can choose to not remove a dead substream for now, and
instead continue processing it at the next iteration. In the example
above, this would make it possible for the dead substream to send its
second event.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>