Report duplicate notification substreams as InboundError events (#2868)
Fixes a TODO.

If a peer opens the same notifications substream multiple times, we
refuse all substreams but the first. This is correct.
However when we refuse the substream we do so completely silently.
Nothing is logged, and thus it makes it hard to figure out whether this
happens.

This PR reports this situation as an event so that the upper layers can
output it in a log for example (that's how `InboundError` events are
handled at the moment).

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>