Remove a hop when sending back JSON-RPC responses (#2851)
Right now, the flow of a JSON-RPC request is like so: it gets pushed to
a queue, then a background task pulls the request, processes it, then
puts back the response in another queue. Then, another background task
pulls the response, and sends it on a channel.

This PR removes the last step (pulling the response and sending it on a
channel). To do so, we completely remove the channel from the public
API, and instead return a new `JsonRpcResponses` object with a `next`
function. This `next` function directly pulls the response from the
queue.

The code is unfortunately more complicated than we could naively think,
because we have to recreate a critical functionality that the channel
provided: when the chain is removed, the channel is closed. Instead we
now have to make the `JsonRpcReponses::next()` function return `None`.
This is done by using another channel whose sole purpose is to detect
when it has been destroyed.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>