Fail if wasm-opt isn't installed when passing `--release` (#2883)
When building the Wasm file to embed in the browser, and `--release` is
provided, we try to run `wasm-opt` on it.
If `--debug` is provided, we don't do it.

Because we used to do it with `--debug` as well, there exists a fallback
if `wasm-opt` isn't installed that will simply not run `wasm-opt`.

This PR removes this fallback, so that the release build fails if
`wasm-opt` isn't installed. `--debug` will still work even if it's not
installed.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>