Reduce the allocations when emitting logs (#2888)
Instead of turning things immediately into a string, we instead put the
values provided to the host function into the `LogEmit` struct, and do
the formatting and turning into string later.

This unlocks a potential future change where the API user can format the
runtime logs the way they want.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>