Remove concept of ready substreams from multi_stream code (#2891)
Right now the state machine of a multi-stream (i.e. WebRTC) connection
has a function returning a list of substreams that have data ready to be
written out.

The objective of this function is simply to allow optimizing the code
that effectively reads/writes, in order to not try to read/write for
nothing.

However this function is currently not properly implemented.
One way to easily implement it better would be to iterate over all
substreams and check which ones are ready. But this is pointless,
because the code that effectively reads/writes can already do that by
trying to read each one one by one.
The best way to implement this function would require a lot of changes
in the lower level substream code, which are quite complicated and I
don't have the motivation to make right now.

This PR simply removes the concept of ready substreams at the moment.
The code simply tries to read/write every single open substream, not
trying to skip the ones that aren't ready.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>