Fix WebRTC for Firefox (#2901)
In the `dataChannel.onmessage` callback, we expect the data to be an
`ArrayBuffer`.
However, in order for that to be the case, one has to set the value of
`binarytype`. See
https://developer.mozilla.org/en-US/docs/Web/API/RTCDataChannel/binaryType.

The documentation says that the default is `blob`. The fact that Chrome
is providing me with an `ArrayBuffer` seems to be a bug:
https://bugs.chromium.org/p/webrtc/issues/detail?id=2276