Add an alternative to RTCCertificate.getFingerprints on Firefox (#2899)
`RTCCertificate.getFingerprints` is famously not available on Firefox.
We can instead generate an SDP offer and parse it.
Unfortunately, this method doesn't work on Chrome, probably because we
generate the SDP offer before creating any channel.
We could change the code to create a channel before generating the
offer, but that would be very hacky and I'd rather avoid this.
Instead, this code has both methods: `getFingerprints` when it's
available, and parsing the SDP offer when it's not.

I've manually checked whether the extraction works by comparing it with
what is in the `about:webrtc` tab, and it does match.

~~After this PR, Firefox no longer throws errors but still doesn't
manage to connect to the server. I don't really know why.~~

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>