Verify that SDP offer contains UDP (#2922)
The comment concerning `m=application` at the bottom says that we tweak
the local offer. This isn't true anymore, but I think it's a good idea
to make sure that the local offer does contain a UDP + data channel
entry.

Tested on Firefox and Chrome. Firefox somehow doesn't work anymore (ugh)
but this PR doesn't actually modify any behavior, so it can't be
related.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>