Print a warning if a legacy JSON-RPC function is used (#2907)
cc https://github.com/paritytech/smoldot/issues/2804

Most legacy JSON-RPC functions can't be implemented properly by smoldot.

The situation kind of sucks, as the new JSON-RPC functions aren't stable
and aren't implemented in Substrate yet.

However, the fact that the replacement isn't ready yet doesn't remove
the fact that the old ones don't work properly. Therefore I think it's
appropriate to print this warning.

The main reason why I'd like to print this warning is to raise awareness
to anyone reading the logs of the fact that things might not work
properly, and that it's not necessarily smoldot's fault but usually the
API's fault.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>