Install clang before the spellcheck (#2935)
Our spellcheck CI job has started failing yesterday because compiling
`clang-sys` can't find LLVM installed on the system.

I was curious why this suddenly happened. Unfortunately, we were caching
the compiled version of `cargo-spellcheck`, and suddenly the cache
expired, which means that we're now compiling it whereas we weren't
compiling it before. I don't really have the motivation to find the CI
run where we initially compiled it in order to compare it with the
failing CI run, and even if I found it the logs are probably expired
now.

A quick investigation of the cargo-spellcheck and clang-sys packages
don't reveal anything abnormal.