Rename handshake module and functions to provide clarity (#2929)
Since multi-stream connections have been introduced, the `handshake`
module doesn't have a great name, as it only concerns single-stream
connections handshakes.

This PR renames it to `single_stream_handshake`.

Additionally, the `new` function is now named `noise_yamux` in order to
make it clear what the handshake does. This provides consistency with
the fact that the WebRTC connection is created with a `webrtc` function.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>