Fix prefix_proof misbehavior (#2947)
Fix https://github.com/paritytech/smoldot/issues/2946

The behavior of `prefix_proof` is currently buggy. When a node isn't
found in the proof that the remote sent us, we simply discard it instead
of adding it to the list of keys to query at the next iteration.
The consequence is that `state_getKeysPaged` was wrong most the time.

In order to fix this, I had to reorganize the entire function.

There's still an issue because we're asking for too much data, which
leads the remote to shutdown the connection, so queries still
occasionally fail. But this other issue is not the fault of the
`prefix_proof` module, and this PR is only about fixing `prefix_proof`.