Make signature verification customizable in the host module (#2955)
cc @xlc 
After looking at https://github.com/paritytech/smoldot/pull/2952, I
thought that the approach in this PR is a better idea.

As shown in the changes to `runtime_host.rs` and
`read_only_runtime_host.rs`, you now have a new variant called
`SignatureVerification` that allows you to overwrite the verification if
you so desire.

As I was implemented this PR, I noticed that the implementation of
`ext_crypto_ecdsa_verify` loads the message from the first parameter,
which is actually the signature. Oops. This is fixed at the same time.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>