Remove key_as_vec() functions (#2957)
When the executor wants to obtain the value at a certain storage
location, it provides us with a `StorageGet` object. This object has a
`key()` function that returns the key to query.

Currently, this function returns an `impl Iterator<Item = impl
AsRef<[u8]>>`, and the actual key consists in the concatenation of all
the elements returned by the iterator.
In order to be more convenient to use, there's also a `key_as_vec()`
function that returns a `Vec<u8>` with the key.

This iterator thing is a micro-optimization that only comes useful at a
single place where we return a list of bytes instead of collecting them
into a `Vec`.
However, the iterator is actually never really usable, and every use of
this API uses `key_as_vec()`, which completely kills the point of this
micro-optimization.
It's also not obvious to me that this optimization actually is an
optimization.

In order to simplify everything, this PR removes the `key_as_vec()`
functions and makes `key()` return a single `impl AsRef<[u8]>` of the
key.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>