Change the state_request protocol to use the proof version (#2966)
At the moment, the state_request protocol code in smoldot uses the "no
proof" version of the protocol. This turns out to be completely useless
in practice. The "no proof" version is supposed to be used only if the
node you query is trusted, which is in practice never the case.

This PR changes that to use the "with proof" version of the protocol,
and adds proper documentation to it.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>