Fix wrong order of parachain blocks at initialization (#2965)
When we subscribe to parachain blocks, the subscription confirmation
contains the list of all blocks that already exist. This list of blocks
needs to be ordered from parent to child, and the field is actually
named `non_finalized_blocks_ancestry_order` ("ancestry order") in order
for everyone to get the memo.

Unfortunately, the author of the parachains code (me) didn't get the
memo, and the parachain blocks aren't properly ordered.
This leads to a panic in `runtime_service.rs` because it assumes that
the ordering is correct.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>