Update to Rust v1.65 and use its new features (#2968)
Updates to Rust 1.65. As usual, the CI will fail because the official
Rust Docker image generally takes a day or two to be updated when a new
Rust version comes out.

This PR introduces GATs in the `Platform` trait.
At the moment, if you call `Platform::next_substream` or
`Platform::next_substream`, you obtain a `Future` that is completely
disconnected from the object representing the connection or the stream.
This means that you can for example drop the connection or stream object
while the `Future` is still alive, or you can call this function
multiple times with the same connection/stream.
Using GATs allows us to bind the `Future` to the object it refers to,
and thus removes these annoying corner cases. This should also make the
trait implementation much easier to write, but I'm not touching this
complicated code right now.

I've switched some code to use `let-else`, but that's just for fun and
this feature is in practice not super useful.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>