Don't call connection_stream_open if WebRTC connection has been reset (#2978)
Fix https://github.com/paritytech/smoldot/issues/2970

The browser can reset a WebRTC connection at any time. When that
happens, smoldot only switches the connection state to "reset", and it's
only later that the connection task will query the state and see that
it's reset.
However, in the meanwhile, the connection task can try to open a new
substream. This PR clarifies the behavior when that happens: trying to
open a new substream should simply be ignored.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>