Return a ValidChainInformation when building the chain spec info (#2979)
The "chain information building process" is done in two steps: fetching
all the info, then validating that the values are coherent together.

In the past, the `chain_spec` module used to do only the first step,
leaving the second step to the API user of this module. Then I've
changed the implementation to produce a `ValidChainInformation` and turn
it back into an (not-validated) `ChainInformation`.

This PR changes the return type to `ValidChainInformation`, and thus the
API user doesn't need to handle that.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>