Decode Merkle proofs entirely at once (#3013)
cc https://github.com/paritytech/smoldot/issues/2973

This PR rewrites the code that verifies Merkle proofs. Instead of
iterating through the proof while decoding it to find one specific
element, we now fully decode and verify the proof then later find the
elements from that decoding. This means that we no longer need to verify
the proof multiple times.

In terms of API, this means that decoding the proof no longer returns
any "entry missing" error. The errors about a missing entry have thus
been added at a higher level.

The logic of the code keeps track of all the entries found in the proof.
Later, in order to find an element, we actually iterate "upwards". So
for example to find `0xf00bab`, we first try `0xf00bab`, then for
example `0xf00b` and then `0xf00`. The vast majority of the time there
should be a precise hit, so it seems better to me to try to find
`0xf00bab` first and then iterate upward in order to prove that the
requested key doesn't exist or isn't in the proof.

I've removed the existing code. Despite the fact that it works, I think
it's a strictly worse way of doing it than the code in this PR.

As I've reached the end of this PR and my brain is fried, I realized
that using a `BTreeMap` is kind of stupid since the proof is already a
tree, and we should instead just keep a parallel `Vec` adding some
information that avoids having to recalculate hashes when decoding the
tree. However I don't have the courage to do this anymore and I've just
left a `TODO` instead. Since this is purely an implementation detail and
that the complexity would remain the same, it can be changed later
without breaking anything.

I also had in mind to immediately add a `next_key` and a `prefix_key`
function to the proof, but again I don't have the courage to do so now.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>